djangoInnowise/requirements.txt
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
psycopg2-binary==2.9.2
djoser==2.1.0
pytest==7.1.2
celery==5.2.7
redis==4.0.2
flower==1.0.0


